[doc] 글 리스트 페이지 클릭 시 상세 페이지로 라우터 구현 완료
